// src/app/modules/shared/models/employees_department/TBLShamelSCLegalHoliday.ts
export interface TBLShamelSCLegalHoliday {
    serial?: number;
    id?:number;
    duration?: number;
    startdate?: Date;
    enddate?: Date;
    documenttype_id?: number;
    document_number?: string;
    documentdate?: Date;
    enterusername?: string;
    enterdate?: Date;
    entertime?: string;
    modifyusername?: string;
    modifydate?: Date;
    modifytime?: string;
}